requirements.txt
python-dotenv==0.20.0
requests==2.27.1
progress==1.6
terminaltables==3.1.10
